[ode16_install] Fix dependency management of ODE 16
Since the dependencies libtool and automake are not dynamic and needed for the bootstrap of ODE, they have to be installed in a previous step to the bootstrap. For that, the manifest is needed and the bootstrap is moved to the configure step, so that the update step has the chance to finish installing the dependencies.