[ode_install_2] Improve ODE 16 package installation and use git instead of archives
The tar.gz import was failing. It seems like the package was attempted to be extracted before finishing the download.
This failure is avoided by using the git repo, in addition, using git repos is helpfull for instance to create MR or patching

Also, sue the default autoproj package management flags instead of system to boostrap the installation. The previous version was failing to compile using the code checked out from the repo.